% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{adj_to_graph}
\alias{adj_to_graph}
\title{Convert adjancency matrix to graph}
\usage{
adj_to_graph(adj, tol = 1e-06)
}
\arguments{
\item{adj}{adjacency matrix}

\item{tol}{tolerance}
}
\value{
\code{igraph} object
}
\description{
Convert adjancency matrix to graph
}
