% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiclique.R
\name{anderson_2009}
\alias{anderson_2009}
\title{Quasiclique based on Anderson and Chellapilla (2009)}
\source{
Andersen, Reid, and Kumar Chellapilla. "Finding dense subgraphs with size bounds." 
International Workshop on Algorithms and Models for the Web-Graph. Springer, Berlin, Heidelberg, 2009.
}
\usage{
anderson_2009(g)
}
\arguments{
\item{g}{\code{igraph} object}
}
\value{
numeric index subset of \code{1:igraph::vcount(g)}
}
\description{
Quasiclique based on Anderson and Chellapilla (2009)
}
