% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai_test.R
\name{cai_test}
\alias{cai_test}
\title{Cai Test}
\usage{
cai_test(x, y, trials = 100, cores = 1, prob = 1)
}
\arguments{
\item{x}{data matrix}

\item{y}{data matrix}

\item{trials}{number of trials}

\item{cores}{number of cores}

\item{prob}{parameter for robustness}
}
\value{
numeric (p-value)
}
\description{
Cai Test
}
