% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{compute_graph_root_distance}
\alias{compute_graph_root_distance}
\title{Compute the distance between two sets of vertices based on graph root embedding}
\usage{
compute_graph_root_distance(eigenvectors, idx1, idx2, k)
}
\arguments{
\item{eigenvectors}{eigenvectors of adjacency matrix}

\item{idx1}{vector containing indices between 1 and \code{nrow(adj)}}

\item{idx2}{vector containing indices between 1 and \code{nrow(adj)}}

\item{k}{number of positive and negative dimensions}
}
\value{
numeric
}
\description{
Compute the distance between two sets of vertices based on graph root embedding
}
