% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{compute_mst_distance}
\alias{compute_mst_distance}
\title{Compute the distance between two sets of vertices based on the MST}
\usage{
compute_mst_distance(adj, idx1, idx2, k)
}
\arguments{
\item{adj}{adjacency matrix}

\item{idx1}{vector containing indices between 1 and \code{nrow(adj)}}

\item{idx2}{vector containing indices between 1 and \code{nrow(adj)}}

\item{k}{positive integer}
}
\value{
numeric
}
\description{
Specifically, focus on the largest connected component in the graph.
Return the number of edges so each vertex in \code{idx1} reaches its
closest \code{k} elements in \code{idx2}
}
