% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.check_node_uniqeness}
\alias{.check_node_uniqeness}
\title{Checks if all nodes in a hash table are unique}
\usage{
.check_node_uniqeness(hash, len)
}
\arguments{
\item{hash}{a \code{hash} object}

\item{len}{maximum index}
}
\value{
boolean
}
\description{
Checks if all nodes in a hash table are unique
}
