% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.check_pairs}
\alias{.check_pairs}
\title{Check pairs of children}
\usage{
.check_pairs(child1, child2, has, mode = "all", null_alarm = F)
}
\arguments{
\item{child1}{vector}

\item{child2}{vector}

\item{has}{a \code{hash} object}

\item{mode}{string}

\item{null_alarm}{boolean}
}
\value{
boolean
}
\description{
Given two vectors, \code{child1} and \code{child2}, and a hash table of
booleans (\code{has}), check to see if any/all of the elements in \code{child1} with
any of the elements in \code{child2} are \code{TRUE} or \code{FALSE}.
}
\details{
\code{child1} and \code{child2} may be \code{NA}, in which case this function
always returns a \code{TRUE}. Otherwise, \code{child1} and \code{child2}
can be a vector of positive integers with length 1 or 2.

When mode "all" is used, a \code{TRUE} is returned only if all pairs between
\code{child1} and \code{child2} are \code{TRUE}. Mode "or" is when only one
pair needs to be \code{TRUE} to output \code{TRUE}.

\code{null_alarm} is a boolean. If \code{TRUE}, this function will throw
an error if a pair is attempted to be checked, but it is not recorded in
\code{has}.
}
