% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.check_superset}
\alias{.check_superset}
\title{Check superset}
\usage{
.check_superset(elements1, elements2)
}
\arguments{
\item{elements1}{vector of indices}

\item{elements2}{vector of indices}
}
\value{
boolean
}
\description{
Return \code{TRUE} if either \code{elements1} or \code{elements2} is a
subset of another
}
