% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{.compute_all_numerator_bootstrap}
\alias{.compute_all_numerator_bootstrap}
\title{Compute all the numerator bootstraps}
\usage{
.compute_all_numerator_bootstrap(dat_list, noise_list, num_list, idx,
  remaining_idx)
}
\arguments{
\item{dat_list}{list of data matrices}

\item{noise_list}{list of noise vectors}

\item{num_list}{list of covariance matrices}

\item{idx}{vector of the indices of the lower triangle}

\item{remaining_idx}{indices to compute the numerator of}
}
\value{
a list of vectors
}
\description{
Compute all the numerator bootstraps
}
