% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai_test.R
\name{.compute_variance}
\alias{.compute_variance}
\title{Compute the variance of the empirical covariance estimate}
\usage{
.compute_variance(mat, cov_vec, idx)
}
\arguments{
\item{mat}{data matrix}

\item{cov_vec}{vectorized covariance matrix of length \code{idx}}

\item{idx}{vector of the indices of the lower triangle}
}
\value{
vector
}
\description{
Compute the variance of the empirical covariance estimate
}
