% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalModel_form.R
\name{.extract_beta}
\alias{.extract_beta}
\title{Loads the \code{glmnet} or \code{cv.glmnet} object and extracts the relevant
coefficients.}
\usage{
.extract_beta(filename_func, lambda = NA, d, num_primary = d,
  cores = 1)
}
\arguments{
\item{filename_func}{a function to create the file name. it takes in
\code{i}, the dimension index, as a parameter}

\item{lambda}{lambda value}

\item{d}{number of dimensions}

\item{num_primary}{integer from 1 to \code{ncol(dat)} that denotes
the number of genes to consider}

\item{cores}{number of cores}
}
\value{
list of vectors
}
\description{
The object loaded is assumed to be called \code{res}. If \code{res} is
not of class \code{cv.glmnet}, \code{lambda} is used. Otherwise,
\code{lambda} is set to \code{res$lambda.1se}.
}
