% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.initialize_unique}
\alias{.initialize_unique}
\title{Initialize hash table of indices}
\usage{
.initialize_unique(lis, d)
}
\arguments{
\item{lis}{list of indices}

\item{d}{maximum index}
}
\value{
a \code{hash} object of booleans
}
\description{
Initialize hash table of indices
}
