% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.node}
\alias{.node}
\title{Create a \code{node} object}
\usage{
.node(elements, children)
}
\arguments{
\item{elements}{vector}

\item{children}{vector}
}
\value{
a \code{node} object
}
\description{
A \code{node} object contains \code{elements} (the indices) and
\code{children} (the assigned number of the two node names that this
current node is made from). Note that this node name is not stored in this
object, but in how it is referenced (in a hash table, for instance).
}
\details{
\code{children} can be either 2 positive integers (in which case the first number
must be small than the second) or \code{NA}.
}
