% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.pass_threshold}
\alias{.pass_threshold}
\title{Determine if an adjacency matrix passes a threshold}
\usage{
.pass_threshold(adj, threshold)
}
\arguments{
\item{adj}{0-1 symmetrical matrix}

\item{threshold}{Binomial probability}
}
\value{
logical
}
\description{
Compares the number of non-zeros in an adjacency matrix to a threshold
}
