% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.post_fillin}
\alias{.post_fillin}
\title{Post fill-in cliques based on certain indices}
\usage{
.post_fillin(clique_list, threshold, target_idx, adj)
}
\arguments{
\item{clique_list}{list of vectors}

\item{threshold}{numeric}

\item{target_idx}{vector}

\item{adj}{adjacency matrix, of elements that are 0 or 1.}
}
\value{
list of vectors
}
\description{
For a list of vectors of indices (\code{clique_list}), determine
if any of elements in \code{target_idx} (a vector of indices) can be
added such that the augmented vector still passes the threshold
in \code{adj}, the adjacency matrix.
}
\details{
Also, if \code{prune} is \code{TRUE}, remove nodes that are not connected
to more than half the nodes.
}
