% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.prune_clique}
\alias{.prune_clique}
\title{Prune the list of cliques}
\usage{
.prune_clique(clique_list, target_idx, prob1 = 0.975, prob2 = 0.995)
}
\arguments{
\item{clique_list}{list of vectors}

\item{target_idx}{vector}

\item{prob1}{numeric}

\item{prob2}{numeric}
}
\value{
list of vectors
}
\description{
This function does find the list of indices in \code{clique_list}
(a list of vectors for indices) based on the intersection of two events:
1) vectors that are above the \code{prob1} quantile in the number of indices
that overlap with \code{target_idx}, and 2) vectors that are above the
\code{prob2} quantile in its length.
}
