% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{.prune_nodes}
\alias{.prune_nodes}
\title{Remove sparsely connected nodes}
\usage{
.prune_nodes(clique_list, adj, target_idx, threshold = 0.5)
}
\arguments{
\item{clique_list}{list of vectors}

\item{adj}{adjacency graph}

\item{target_idx}{vector of indices from 1 to \code{ncol(adj)} to force vectors in \code{clique_list} to retain}

\item{threshold}{number between 0 and 1}
}
\value{
list of vectors
}
\description{
Remove sparsely connected nodes
}
