% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{.row_norm}
\alias{.row_norm}
\title{Row normalization}
\usage{
.row_norm(A)
}
\arguments{
\item{A}{the input matrix}
}
\value{
the new matrix after row-normalization.
}
\description{
Normalize the non-zero rows of a matrix A
to have unit L2 norm
}
