% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{.spectral_cluster}
\alias{.spectral_cluster}
\title{Spectral clustering}
\usage{
.spectral_cluster(g, K = 2)
}
\arguments{
\item{g}{\code{igraph} object}

\item{K}{number of clusters}
}
\value{
vector
}
\description{
Spectral clustering
}
