% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{enrichment_test}
\alias{enrichment_test}
\title{Fisher test}
\usage{
enrichment_test(vec1, vec2, all_vec)
}
\arguments{
\item{vec1}{vector of gene names}

\item{vec2}{vector of gene names}

\item{all_vec}{vector of all gene names}
}
\value{
list containing the contigency matrix and the Fisher p value
}
\description{
Fisher test
}
