% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{enumerate_edges}
\alias{enumerate_edges}
\title{List the index-pairs of non-zeros in an adjacency matrix}
\usage{
enumerate_edges(adj, tol = 1e-06)
}
\arguments{
\item{adj}{matrix}

\item{tol}{small number}
}
\value{
matrix with 2 columns
}
\description{
Only lists elements in the upper-triange, excluding diagonal
}
