% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\name{filename_closure}
\alias{filename_closure}
\title{Create filename function}
\usage{
filename_closure(folder, prefix = "")
}
\arguments{
\item{folder}{folder name (character)}

\item{prefix}{prefix that all filenames will receive (character)}
}
\value{
function
}
\description{
If \code{folder} does not exist, create \code{folder}. Returns
a function to make the filename. \code{folder} must be a character
that ends with "/".
}
