% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Goodness of fit diagnostic}
\usage{
goodness_of_fit(dat_list, num_partitions, trials = 500, cores = 1,
  verbose = F)
}
\arguments{
\item{dat_list}{list of data matrices with same number of columns}

\item{num_partitions}{number of partitions to attempt}

\item{trials}{number of trials}

\item{cores}{number of cores}

\item{verbose}{boolean}
}
\value{
vector of p-values
}
\description{
Partitions a list of datasets into two halves, and runs the Cai covariance
test on the two halves. This function first centers each half before
aggregating. The function tries different unique ways to partition the data
into two halves, and returns a p-value for each attempted partition.
}
