% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalModel.R
\name{graphicalModel}
\alias{graphicalModel}
\title{Graphical model estimate}
\usage{
graphicalModel(dat, lambda = "lambda.1se", verbose = F, tol = 0.001)
}
\arguments{
\item{dat}{the nxd matrix}

\item{verbose}{boolean}
}
\value{
a precision matrix estimate, dxd
}
\description{
Estimated using neighbhorhood selection, cross validation to select lambda
}
