% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalModel_form.R
\name{graphicalModel_form}
\alias{graphicalModel_form}
\title{Graphical model estimate}
\usage{
graphicalModel_form(dat, filename_func, lambda = NA,
  num_primary = ncol(dat), tol = 1e-06, cores = 1, verbose = F)
}
\arguments{
\item{dat}{the nxd matrix}

\item{filename_func}{a function to create the file name. it takes in
\code{i}, the dimension index, as a parameter}

\item{lambda}{lambda value}

\item{num_primary}{integer from 1 to \code{ncol(dat)} that denotes
the number of genes to consider}

\item{tol}{numeric}

\item{cores}{number of cores}

\item{verbose}{boolean for verbose}
}
\value{
a precision matrix estimate, dxd
}
\description{
After a call of \code{graphicalModel_store}, \code{graphicalmodel_form}
will take the saved glmnet objects (dictated by \code{filename_func})
to form the precision matrix
}
