% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalModel.R
\name{graphicalModel_range}
\alias{graphicalModel_range}
\title{Graphical model estimate for a range of lambda values}
\usage{
graphicalModel_range(dat, primary_idx, lambda_min, lambda_max,
  lambda_length = 15, verbose = F, tol = 1e-06)
}
\arguments{
\item{dat}{the matrix with \code{n} rows and \code{d} columns}

\item{primary_idx}{index vector that is a subset of \code{1:ncol(dat)}}

\item{lambda_min}{minimum value of \code{lambda} when using \code{graphicalModel}}

\item{lambda_max}{maximum value of \code{lambda} when using \code{graphicalModel}}

\item{lambda_length}{number of \code{lambda} values to try (exponential growth)}

\item{verbose}{boolean}

\item{tol}{numeric}
}
\value{
a list, each being an output for \code{graphicalModel} for a different value of \code{lambda}
}
\description{
Graphical model estimate for a range of lambda values
}
