% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalModel_store.R
\name{graphicalModel_store}
\alias{graphicalModel_store}
\title{Graphical model estimate}
\usage{
graphicalModel_store(dat, filename_func, num_primary = ncol(dat), cores = 1,
  cv = F)
}
\arguments{
\item{dat}{the nxd matrix}

\item{filename_func}{a function to create the file name. it takes in
\code{i}, the dimension index, as a paramter}

\item{num_primary}{integer from 1 to \code{ncol(dat)} that denotes
the number of genes to consider}

\item{cores}{number of cores}

\item{cv}{boolean on whether or not to cross validate}
}
\value{
void
}
\description{
Store the glmnet objects (one for each dimension) in a folder location.
Will NOT create the folder, please make sure done beforehand.
This function requires cores to be set up for parallelization.
The objects are saved as name "res".
}
