% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iossifov}
\alias{iossifov}
\title{Iossifov TADA genes}
\format{A vector of characters, length 251}
\source{
The partitionings were based on the data used in
"Iossifov, I., ORoak, B. J., Sanders, S. J., Ronemus, M., Krumm, N., Levy,
D., Stessman, H. A., Witherspoon, K. T., Vives, L., and Patterson, K. E. (2014).
The contribution of de novo coding mutations to autism spectrum disorder.
Nature, 515(7526):216–221.
}
\description{
Iossifov TADA genes
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
