% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_subregion}
\alias{region_subregion}
\title{Partitioning of the regions in the brain}
\format{A data frame with two columns
\describe{
  \item{region}{region}
  \item{subregion}{subregion}
}}
\source{
The partitionings were based on
"Willsey, A. J., Sanders, S. J., Li, M., Dong, S., Tebbenkamp, A. T., Muhle,
R. A., Reilly, S. K., Lin, L., Fertuzinhos, S., and Miller, J. A. (2013).
Coexpression networks implicate human midfetal deep cortical projection
neurons in the pathogenesis of autism. Cell, 155(5):997–1007."
}
\description{
Partitioning of the regions in the brain
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
