% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{regroup}
\alias{regroup}
\title{Regroup a list of matrices into another list of matrices}
\usage{
regroup(dat_list, idx_list)
}
\arguments{
\item{dat_list}{a list of matrices}

\item{idx_list}{a list of integer vectors}
}
\value{
a list of matrices
}
\description{
\code{dat_list} is a list of matrices, and \code{idx_list} is
a list of integer vectors, where each integer from 1 to \code{length(dat_list)}
appears exactly once. Outputs a list of matrices of length \code{length(idx_list)}
matrices.
}
