% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{report_results}
\alias{report_results}
\title{Report results}
\usage{
report_results(gene, posterior, pvalue, Iupdate)
}
\arguments{
\item{gene}{vector of characters}

\item{posterior}{vector of numerics}

\item{pvalue}{vector of numerics}

\item{Iupdate}{vector of booleans}
}
\value{
data frame with 4 columns
}
\description{
Report results
}
