% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen}
\alias{screen}
\title{Screening data}
\usage{
screen(dat, pv, p_thres = 0.1, cor_thres = 0.7)
}
\arguments{
\item{dat}{the nxd matrix}

\item{pv}{vector of d pvalues}

\item{p_thres}{numeric}

\item{cor_thres}{numeric}
}
\value{
list of primary and secondary genes, each a vector of integers
between 1 and \code{ncol(dat)}
}
\description{
We use a screening method where genes with \code{pv} lower than
\code{p_thres} are primary genes, and the secondary genes are selected based on having the having
(absolute) Pearson correlation with any of the primary genes
higher than \code{cor_thres}
}
