% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{stepdown}
\alias{stepdown}
\title{Stepdown to test all pairs of datasets to see if they have equal covariance}
\usage{
stepdown(dat_list, trials = 100, alpha = 0.05, return_pvalue = F,
  cores = 1, verbose = F)
}
\arguments{
\item{dat_list}{list of data matrices with same number of columns}

\item{trials}{number of trials}

\item{alpha}{alpha level}

\item{cores}{number of cores}

\item{verbose}{boolean for verbose}

\item{denominator}{boolean}
}
\value{
indices for \code{combn(length(dat_list), 2)} that correspond to the
hypotheses that passed
}
\description{
Stepdown to test all pairs of datasets to see if they have equal covariance
}
