% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_path.R
\name{stepdown_choose}
\alias{stepdown_choose}
\title{Stepdown choose}
\usage{
stepdown_choose(stepdown_obj, alpha = 0.05, return_pvalue = F,
  verbose = F)
}
\arguments{
\item{stepdown_obj}{stepdown object}

\item{alpha}{alpha level}

\item{return_pvalue}{boolean if the naive p-values should be returned}

\item{verbose}{boolean for verbose}
}
\value{
indices for \code{combn(length(dat_list), 2)} that correspond to the
hypotheses that passed
}
\description{
Stepdown choose
}
