% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_path.R
\name{stepdown_path}
\alias{stepdown_path}
\title{Stepdown path}
\usage{
stepdown_path(dat_list, trials = 100, iterations = 15, cores = 1,
  verbose = F, file = NA, prob = 1, squared = T)
}
\arguments{
\item{dat_list}{list of data matrices with same number of columns}

\item{trials}{number of trials}

\item{iterations}{number of iterations}

\item{cores}{number of cores}

\item{verbose}{boolean for verbose}

\item{file}{temporary save filepath}

\item{prob}{numeric between 0 and 1}
}
\value{
a \code{stepdown} object
}
\description{
Stepdown path
}
