% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synonyms}
\alias{synonyms}
\title{Gene synonyms}
\format{A list
\describe{
  \item{hash}{A \code{hash} object where each key is a symbol and each value is an integer.
  The integer references which element in \code{syn.list} corresponds to the key.}
  \item{syn.list}{A list where each element (named) is a vector of synonym symbols (characters)}
}}
\source{
The dataset is based on outputs in \url{http://biomart.genenames.org/martform/#!/default/HGNC?datasets=hgnc_gene_mart}.
}
\description{
Gene synonyms
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
