% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tada}
\alias{tada}
\title{TADA dataset}
\format{A data frame with 18735 rows (18735 genes) and 17 columns.
\describe{
  \item{Gene}{Gene name}
  \item{pval.TADA}{pvalue}
}}
\source{
The methodology to calculate these values were based on
"He, X., Sanders, S. J., Liu, L., De Rubeis, S., Lim, E. T., Sutcliffe, J. S.,
... & Devlin, B. (2013). Integrated model of de novo and inherited genetic
variants yields greater power to identify risk genes. PLoS genetics, 9(8), e1003671.",
but the dataset itself comes from 
"De Rubeis, Silvia, et al. "Synaptic, transcriptional and chromatin genes disrupted in autism." Nature 515.7526 (2014): 209."
}
\description{
TADA dataset
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
