% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validated_genes}
\alias{validated_genes}
\title{Partitioning of the regions in the brain}
\format{A data frame with one column
\describe{
  \item{Gene}{gene name}
}}
\source{
Genes based on the findings of
"Satterstrom, F. Kyle, et al. "Large-scale exome sequencing study implicates both developmental 
and functional changes in the neurobiology of autism." (2019)."
}
\description{
Partitioning of the regions in the brain
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
