% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tada}
\alias{tada}
\title{TADA dataset}
\format{A data frame with 18735 rows (18735 genes) and 17 columns.
\describe{
  \item{Gene}{Gene name}
  \item{mut.rate}{mutation rate}
  \item{dn.LoF}{number of denovo loss of function events}
  \item{case.LoF}{number of loss of function events in case}
  \item{ctrl.LoF}{number of loss of function events in control}
  \item{trans.LoF}{number of transmitted loss of function events}
  \item{ntrans.LoF}{number of nontransmitted loss of function events}
  \item{dn.mis3}{number of denovo missense 3 mutation events}
  \item{case.mis3}{number of missense 3 mutation events in case}
  \item{ctrl.mis3}{number of missense 3 mutation events in control}
  \item{trans.mis3}{number of transmitted missense 3 events}
  \item{ntrans.mis3}{number of nontransmitted missense 3 events}
  \item{BF.dn}{Bayes factor for denovo}
  \item{BF}{Bayes factor}
  \item{qvalue.dn}{q value for denovo}
  \item{qvalue}{qvalue}
  \item{pval.TADA}{pvalue}
}}
\source{
The methodology to calculate these values were based on
"He, X., Sanders, S. J., Liu, L., De Rubeis, S., Lim, E. T., Sutcliffe, J. S.,
... & Devlin, B. (2013). Integrated model of de novo and inherited genetic
variants yields greater power to identify risk genes. PLoS genetics, 9(8), e1003671.",
but the dataset itself comes from 
"De Rubeis, Silvia, et al. "Synaptic, transcriptional and chromatin genes disrupted in autism." Nature 515.7526 (2014): 209."
}
\description{
TADA dataset
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
